Fix encoded_frame_test
